resolab-backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
boto3==1.16.34
botocore==1.19.34
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
coreapi==2.3.3
coreschema==0.0.4
crayons==0.3.0
cryptography==3.2.1
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.5.0
django-filter==2.4.0
django-utils==0.0.2
djangorestframework==3.12.2
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
firebase==3.0.1
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
inflection==0.5.1
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.10.0
json5==0.9.0
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
lxml==4.4.2
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
packaging==20.7
pandocfilters==1.4.2
parso==0.6.0
pickleshare==0.7.5
Pillow==7.1.2
pipenv==2020.11.15
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.13.0
psycopg2==2.8.5
pyAesCrypt==0.4.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.5.2
pyparsing==2.4.7
pyrsistent==0.15.7
python-dateutil==2.8.1
python-decouple==3.3
python-docx==0.8.10
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
razorpay==1.2.0
requests==2.23.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.19
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
waitress==1.4.4
wcwidth==0.1.8
webdriver-manager==2.3.0
webencodings==0.5.1
wget==3.2
whitenoise==5.2.0
